+++ bindings/python/src/tokenizer.rs
@@ -294,17 +294,17 @@ impl Tokenizer {
 
     #[args(kwargs = "**")]
     fn enable_truncation(&mut self, max_length: usize, kwargs: Option<&PyDict>) -> PyResult<()> {
-        let mut stride = 0;
-        let mut strategy = TruncationStrategy::LongestFirst;
+        let mut params = TruncationParams::default();
+        params.max_length = max_length;
 
         if let Some(kwargs) = kwargs {
             for (key, value) in kwargs {
                 let key: &str = key.extract()?;
                 match key {
-                    "stride" => stride = value.extract()?,
+                    "stride" => params.stride = value.extract()?,
                     "strategy" => {
                         let value: &str = value.extract()?;
-                        strategy = match value {
+                        params.strategy = match value {
                             "longest_first" => Ok(TruncationStrategy::LongestFirst),
                             "only_first" => Ok(TruncationStrategy::OnlyFirst),
                             "only_second" => Ok(TruncationStrategy::OnlySecond),
@@ -321,11 +321,7 @@ impl Tokenizer {
             }
         }
 
-        self.tokenizer.with_truncation(Some(TruncationParams {
-            max_length,
-            stride,
-            strategy,
-        }));
+        self.tokenizer.with_truncation(Some(params));
 
         Ok(())
     }
@@ -334,14 +330,22 @@ impl Tokenizer {
         self.tokenizer.with_truncation(None);
     }
 
+    #[getter]
+    fn get_truncation<'py>(&self, py: Python<'py>) -> PyResult<Option<&'py PyDict>> {
+        self.tokenizer.get_truncation().map_or(Ok(None), |params| {
+            let dict = PyDict::new(py);
+
+            dict.set_item("max_length", params.max_length)?;
+            dict.set_item("stride", params.stride)?;
+            dict.set_item("strategy", params.strategy.as_ref())?;
+
+            Ok(Some(dict))
+        })
+    }
+
     #[args(kwargs = "**")]
     fn enable_padding(&mut self, kwargs: Option<&PyDict>) -> PyResult<()> {
-        let mut direction = PaddingDirection::Right;
-        let mut pad_to_multiple_of: Option<usize> = None;
-        let mut pad_id: u32 = 0;
-        let mut pad_type_id: u32 = 0;
-        let mut pad_token = String::from("[PAD]");
-        let mut max_length: Option<usize> = None;
+        let mut params = PaddingParams::default();
 
         if let Some(kwargs) = kwargs {
             for (key, value) in kwargs {
@@ -349,7 +353,7 @@ impl Tokenizer {
                 match key {
                     "direction" => {
                         let value: &str = value.extract()?;
-                        direction = match value {
+                        params.direction = match value {
                             "left" => Ok(PaddingDirection::Left),
                             "right" => Ok(PaddingDirection::Right),
                             other => Err(PyError(format!(
@@ -360,36 +364,66 @@ impl Tokenizer {
                             .into_pyerr()),
                         }?;
                     }
-                    "pad_to_multiple_of" => pad_to_multiple_of = value.extract()?,
-                    "pad_id" => pad_id = value.extract()?,
-                    "pad_type_id" => pad_type_id = value.extract()?,
-                    "pad_token" => pad_token = value.extract()?,
-                    "max_length" => max_length = value.extract()?,
+                    "pad_to_multiple_of" => {
+                        if let Some(multiple) = value.extract()? {
+                            params.pad_to_multiple_of = multiple;
+                        }
+                    }
+                    "pad_id" => params.pad_id = value.extract()?,
+                    "pad_type_id" => params.pad_type_id = value.extract()?,
+                    "pad_token" => params.pad_token = value.extract()?,
+                    "max_length" => {
+                        println!(
+                            "enable_padding(max_length=X) is deprecated, \
+                                 use enable_padding(length=X) instead"
+                        );
+                        if let Some(l) = value.extract()? {
+                            params.strategy = PaddingStrategy::Fixed(l);
+                        } else {
+                            params.strategy = PaddingStrategy::BatchLongest;
+                        }
+                    }
+                    "length" => {
+                        if let Some(l) = value.extract()? {
+                            params.strategy = PaddingStrategy::Fixed(l);
+                        } else {
+                            params.strategy = PaddingStrategy::BatchLongest;
+                        }
+                    }
                     _ => println!("Ignored unknown kwarg option {}", key),
                 }
             }
         }
 
-        let strategy = if let Some(max_length) = max_length {
-            PaddingStrategy::Fixed(max_length)
-        } else {
-            PaddingStrategy::BatchLongest
-        };
-
-        self.tokenizer.with_padding(Some(PaddingParams {
-            strategy,
-            direction,
-            pad_to_multiple_of,
-            pad_id,
-            pad_type_id,
-            pad_token: pad_token.to_owned(),
-        }));
+        self.tokenizer.with_padding(Some(params));
 
         Ok(())
     }
 
     fn no_padding(&mut self) {
         self.tokenizer.with_padding(None);
+    }
+
+    #[getter]
+    fn get_padding<'py>(&self, py: Python<'py>) -> PyResult<Option<&'py PyDict>> {
+        self.tokenizer.get_padding().map_or(Ok(None), |params| {
+            let dict = PyDict::new(py);
+
+            dict.set_item(
+                "length",
+                match params.strategy {
+                    tk::PaddingStrategy::BatchLongest => None,
+                    tk::PaddingStrategy::Fixed(size) => Some(size),
+                },
+            )?;
+            dict.set_item("pad_to_multiple_of", params.pad_to_multiple_of)?;
+            dict.set_item("pad_id", params.pad_id)?;
+            dict.set_item("pad_token", &params.pad_token)?;
+            dict.set_item("pad_type_id", params.pad_type_id)?;
+            dict.set_item("direction", params.direction.as_ref())?;
+
+            Ok(Some(dict))
+        })
     }
 
     fn normalize(&self, sentence: &str) -> PyResult<String> {